Add soma clustering from simulation-templates repo.
