+++ demo/soma_clustering/CMakeLists.txt
@@ -17,13 +17,13 @@ cmake_minimum_required(VERSION 3.2.0)
 project(soma_clustering)
 
 find_package(BioDynaMo REQUIRED)
-include("${BDM_USE_FILE}")
+include(${BDM_USE_FILE})
 include_directories("src")
 
 file(GLOB_RECURSE HEADERS src/*.h)
 file(GLOB_RECURSE SOURCES src/*.cc)
 
 bdm_add_executable(soma_clustering
-                   HEADERS "${HEADERS}"
-                   SOURCES "${SOURCES}"
-                   LIBRARIES "${BDM_REQUIRED_LIBRARIES}")
+                   HEADERS ${HEADERS}
+                   SOURCES ${SOURCES}
+                   LIBRARIES ${BDM_REQUIRED_LIBRARIES})
